Update README.md (initialisation timing)
Hi all, this is more of a trial of a pull request rather than a serious issue...
Shall we edit the README to reflect the fact that the population initialisation takes (or can take) longer than 10 minutes?
At least on my computer (which I think is rather strong) it takes around 2 hours to run the population initialisation step, I have run it few times already and had always the same performance. Happy to discuss at the stand-up. Thanks!